+++ packages/moonstone/Panels/Panel.js
@@ -1,3 +1,4 @@
+import deprecate from '@enact/core/internal/deprecate';
 import {forward, handle} from '@enact/core/handle';
 import kind from '@enact/core/kind';
 import React from 'react';
@@ -5,26 +6,38 @@ import PropTypes from 'prop-types';
 import Slottable from '@enact/ui/Slottable';
 import Spotlight from '@enact/spotlight';
 import SpotlightContainerDecorator from '@enact/spotlight/SpotlightContainerDecorator';
-import {spottableClass} from '@enact/spotlight/Spottable';
 
 import css from './Panel.less';
 
-const spotPanel = (node) => {
+const spotPanel = (autoFocus) => (node) => {
 	if (node && !Spotlight.getCurrent()) {
 		const {containerId} = node.dataset;
+		const config = {
+			enterTo: 'last-focused'
+		};
 
-		// set the default element so that we can try to spot the container and let it fall through
-		// to the default element if there isn't a lastFocusedElement/Index
-		const body = node.querySelector(`section .${spottableClass}`);
-		const header = node.querySelector(`header .${spottableClass}`);
-		const defaultElement = body || header;
+		if (autoFocus !== 'last-focused') {
+			config.enterTo = 'default-element';
 
-		Spotlight.set(containerId, {defaultElement});
+			if (autoFocus !== 'default-element') {
+				config.defaultElement = autoFocus;
+			}
+		}
+
+		Spotlight.set(containerId, config);
 		Spotlight.focus(containerId);
 	}
 };
 
 let panelId = 0;
+
+// Called when `noAutoFocus` is true. Warns the first time and returns the new `autoFocus` value
+const adaptToAutoFocus = deprecate(() => 'none', {
+	name: 'noAutoFocus',
+	since: '1.3.0',
+	until: '2.0.0',
+	replacedBy: 'autoFocus'
+});
 
 /**
 * {@link moonstone/Panels.Panel} is the default kind for controls created inside a
@@ -53,6 +66,21 @@ const PanelBase = kind({
 		'aria-label': PropTypes.string,
 
 		/**
+		 * Sets the strategy used to automatically focus an element within the panel upon render.
+		 *
+		 * * "none" - Automatic focus is disabled
+		 * * "last-focused" - The element last focused in the panel with be restored
+		 * * "default-element" - The first spottable component within the body will be focused
+		 * * Custom Selector - A custom CSS selector may also be provided which will be used to find
+		 *   the target within the Panel
+		 *
+		 * @type {String}
+		 * @default 'last-focused'
+		 * @public
+		 */
+		autoFocus: PropTypes.string,
+
+		/**
 		 * Header for the panel. This is usually passed by the {@link ui/Slottable.Slottable} API by
 		 * using a [Header]{@link moonstone/Panels.Header} component as a child of the Panel.
 		 *
@@ -77,8 +105,10 @@ const PanelBase = kind({
 		hideChildren: PropTypes.bool,
 
 		/**
-		 * When `true`, the contents of the Panel will not receive spotlight focus after being rendered.
+		 * When `true`, the contents of the Panel will not receive spotlight focus after being
+		 * rendered.
 		 *
+		 * @deprecated Replaced by `autoFocus="none"`
 		 * @type {Boolean}
 		 * @default false
 		 * @public
@@ -87,6 +117,7 @@ const PanelBase = kind({
 	},
 
 	defaultProps: {
+		autoFocus: 'last-focused',
 		hideChildren: false,
 		noAutoFocus: false
 	},
@@ -110,7 +141,17 @@ const PanelBase = kind({
 		// In order to spot the body components, we defer spotting until !hideChildren. If the Panel
 		// opts out of hideChildren support by explicitly setting it to false, it'll spot on first
 		// render.
-		spotOnRender: ({hideChildren, noAutoFocus}) => hideChildren || noAutoFocus ? null : spotPanel,
+		spotOnRender: ({autoFocus, hideChildren, noAutoFocus}) => {
+			if (noAutoFocus) {
+				autoFocus = adaptToAutoFocus();
+			}
+
+			if (hideChildren || autoFocus === 'none') {
+				return null;
+			}
+
+			return spotPanel(autoFocus);
+		},
 		children: ({children, hideChildren}) => hideChildren ? null : children,
 		bodyClassName: ({header, hideChildren, styler}) => styler.join({
 			body: true,
@@ -124,6 +165,7 @@ const PanelBase = kind({
 	},
 
 	render: ({bodyClassName, children, header, headerId, spotOnRender, ...rest}) => {
+		delete rest.autoFocus;
 		delete rest.hideChildren;
 		delete rest.noAutoFocus;
 
@@ -137,7 +179,12 @@ const PanelBase = kind({
 });
 
 const Panel = SpotlightContainerDecorator(
-	{enterTo: 'last-focused', preserveId: true},
+	{
+		// prefer any spottable within the panel body for first render
+		defaultElement: `.${css.body} *`,
+		enterTo: 'last-focused',
+		preserveId: true
+	},
 	Slottable(
 		{slots: ['header']},
 		PanelBase